+++ client/src/commands.ts
@@ -164,17 +164,21 @@ export function startLanguageServer(
     env["NO_COLOR"] = "1";
     env["DENO_V8_FLAGS"] = getV8Flags();
 
+    const lspPath = path.resolve(context.extensionPath, 'client', 'lsp-server', 'notebook-lsp.ts');
+
     const serverOptions: ServerOptions = {
       run: {
         command,
-        args: ["lsp"],
+        //args: ["lsp"],
+        args: ['--allow-env', '--allow-run', lspPath, JSON.stringify(env)],
         options: { env },
       },
       debug: {
         command,
+        //args: ["lsp"],        
+        args: ['--allow-env', '--allow-run', lspPath, JSON.stringify(env)],
         // disabled for now, as this gets super chatty during development
         // args: ["lsp", "-L", "debug"],
-        args: ["lsp"],
         options: { env },
       },
     };
@@ -654,6 +658,7 @@ export function enable(
   return async () => {
     const config = vscode.workspace.getConfiguration(EXTENSION_NS);
     await config.update("enable", true);
+    await vscode.workspace.getConfiguration('typescript.validate').update('enable', false);
     vscode.window.showInformationMessage("Deno workspace initialized.");
     const tsserverConfig = vscode.workspace.getConfiguration(
       "typescript.tsserver",
@@ -684,6 +689,7 @@ export function disable(
   return async () => {
     const config = vscode.workspace.getConfiguration(EXTENSION_NS);
     await config.update("enable", false);
+    await vscode.workspace.getConfiguration('typescript.validate').update('enable', true);
   };
 }